+++ R/app_version.R
@@ -1,1 +1,1 @@
-app_version <- function()'26.3.1'
+app_version <- function()'26.3.2'