+++ R/hosted_version.R
@@ -1,1 +1,1 @@
-hosted_version <- function()'v26.3.1-260302'
+hosted_version <- function()'v26.3.2-260311'